+++ h3ron-graph/src/algorithm/mod.rs
@@ -3,3 +3,4 @@ pub mod differential_shortest_path;
 mod dijkstra;
 pub mod path;
 pub mod shortest_path;
+pub mod within_weight_threshold;